% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_rm_grp_NA.R
\name{rm_grp_na}
\alias{rm_grp_na}
\title{Remove groups where all values in a target column are NA}
\usage{
rm_grp_na(df, group_vars, target_col)
}
\arguments{
\item{df}{A data frame or tibble}

\item{group_vars}{A character vector of column names to group by}

\item{target_col}{A character string naming the column to check for NA}
}
\value{
A filtered data frame where any group with all NAs in \code{target_col} is removed
}
\description{
Remove groups where all values in a target column are NA
}
\examples{
df <- tibble(
  id = c(1, 1, 2, 2, 3, 3),
  time = c("A", "B", "A", "B", "A", "B"),
  value = c(NA, NA, 10, NA, 5, 6)
)

# Remove groups where all `value` entries are NA, grouped by `id`
remove_all_na_groups(df, group_vars = c("id"), target_col = "value")

}
